requirements.txt
certifi==2020.12.5
chardet==4.0.0
fastapi==0.63.0
geographiclib==1.50
geopy==2.0.0
idna==2.10
numpy==1.19.4
pandas==1.2.0
pydantic==1.7.3
pyodbc==4.0.30
pyproj==3.0.0.post1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
Shapely==1.7.1
six==1.15.0
starlette==0.13.6
urllib3==1.26.2
